Add form class && some docs && delete .env
