--- .env.sample
@@ -1,2 +1,0 @@
-VITE_NOTION_DATABASE_ID=""
-VITE_NOTION_CLIENT_SECRET=""
