--- src/services/Config.ts
@@ -1,6 +1,0 @@
-export type NotionConfig = {
-  notion: {
-    databaseId: string;
-    clientSecret: string;
-  };
-};
